+++ src/main/sarl/fr/utbm/info/vi51/worldswar/agent/tactical_behaviour/GoHomeTacticalBehaviour.java
@@ -3,8 +3,8 @@ package fr.utbm.info.vi51.worldswar.agent.tactical_behaviour;
 import java.util.HashMap;
 
 import fr.utbm.info.vi51.worldswar.agent.operational_behaviour.MoveToTargetOperationalBehaviour;
-import fr.utbm.info.vi51.worldswar.environment.PerceptionGrid;
 import fr.utbm.info.vi51.worldswar.environment.influence.Influence;
+import fr.utbm.info.vi51.worldswar.perception.PerceptionGrid;
 
 public class GoHomeTacticalBehaviour implements TacticalBehaviour {
