+++ src/main/sarl/fr/utbm/info/vi51/worldswar/agent/tactical_behaviour/TacticalBehaviour.java
@@ -2,8 +2,8 @@ package fr.utbm.info.vi51.worldswar.agent.tactical_behaviour;
 
 import java.util.HashMap;
 
-import fr.utbm.info.vi51.worldswar.environment.PerceptionGrid;
 import fr.utbm.info.vi51.worldswar.environment.influence.Influence;
+import fr.utbm.info.vi51.worldswar.perception.PerceptionGrid;
 
 public interface TacticalBehaviour {
 	Influence computeInfluence(PerceptionGrid perceptionGrid, HashMap<String, Object> memory);
