+++ src/main/sarl/fr/utbm/info/vi51/worldswar/environment/Environment.sarl
@@ -16,6 +16,7 @@ import fr.utbm.info.vi51.worldswar.environment.influence.MeleeAttackInfluence
 import fr.utbm.info.vi51.worldswar.environment.influence.MoveInfluence
 import fr.utbm.info.vi51.worldswar.environment.influence.PheromoneAndMoveInfluence
 import fr.utbm.info.vi51.worldswar.environment.influence.PickFoodInfluence
+import fr.utbm.info.vi51.worldswar.perception.PerceptionGrid
 import fr.utbm.info.vi51.worldswar.simulator.SimulationStepEvent
 import fr.utbm.info.vi51.worldswar.simulator.SimulatorListener
 import fr.utbm.info.vi51.worldswar.simulator.StopSimulationEvent
