+++ src/main/sarl/fr/utbm/info/vi51/worldswar/simulator/SimulatorListener.java
@@ -1,6 +1,6 @@
 package fr.utbm.info.vi51.worldswar.simulator;
 
-import fr.utbm.info.vi51.worldswar.environment.PerceptionGrid;
+import fr.utbm.info.vi51.worldswar.perception.PerceptionGrid;
 
 /**
  * Listener interface about a {@link Simulator}
